fb599: renaming to administration-service
